refactor: move middleware under src/http/
